custom exceptions added, new functions on controllers, and new sql data

+++ app/Http/Controllers/CustomQuestionController.php
@@ -2,35 +2,35 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\CustomNotFound;
+use App\Exceptions\SQLInfraction;
 use App\Models\CustomQuestion;
 use Illuminate\Http\Request;
+use Mockery\Exception;
 
 class CustomQuestionController extends Controller
 {
-    public function allCustomQuestions(): \Illuminate\Http\JsonResponse
-    {
-        $results = CustomQuestion::all();
-        return response()->json($results);
-    }
 
 
     public function createCustomQuestion(Request $request)
     {
         $question = CustomQuestion::create($request->all());
+        if (!$question) throw new Exception('not created', 500);
         return response()->json($question, 201);
     }
 
     public function getAllCustomQuestionsFromCustomQuiz($id_quiz)
     {
-        $question = CustomQuestion::where('fk_id_quiz', $id_quiz)->get();
-        return response()->json($question, 201);
+        $questions = CustomQuestion::where('fk_id_quiz', $id_quiz)->get();
+        if (empty($questions)) throw new CustomNotFound('not questions found');
+        return response()->json($questions, 201);
     }
 
 
     public function updateCustomQuestion(Request $request, $id)
     {
         $question = CustomQuestion::findOrFail($id);
-
+        if (!$question) throw new CustomNotFound('not questions found');
         $request->validate(CustomQuestion::rules());
 
         $question->fill($request->only([
@@ -49,6 +49,7 @@ class CustomQuestionController extends Controller
     public function deleteCustomQuestion($id)
     {
         $question = CustomQuestion::findOrFail($id);
+        if (!$question) throw new CustomNotFound('not questions found');
         $question->delete();
         return response()->json(null, 204);
     }
