+++ app/Http/Controllers/CustomQuizController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\CustomNotFound;
 use App\Models\CustomQuiz;
 use Illuminate\Http\Request;
 
@@ -11,19 +12,29 @@ class CustomQuizController extends Controller
     public function allCustoms()
     {
         $customQuizzes = CustomQuiz::all();
+        if (!$customQuizzes) throw new CustomNotFound('no quizzes found');
+        return response()->json($customQuizzes);
+    }
+
+    public function allCustomsByUser($id_user)
+    {
+        $customQuizzes = CustomQuiz::where('fk_id_user', $id_user)->get();
+        if (!$customQuizzes) throw new CustomNotFound('no quizzes found for the user');
         return response()->json($customQuizzes);
     }
 
     public function createCustom(Request $request)
     {
         $customQuiz = CustomQuiz::create($request->all());
+        if (!$customQuiz) throw new \Mockery\Exception('not created', 500);
         $request->validate(CustomQuiz::rules());
         return response()->json($customQuiz, 201);
     }
 
-    public function getCustom($id)
+    public function getCustom($id_custom)
     {
-        $customQuiz = CustomQuiz::findOrFail($id);
+        $customQuiz = CustomQuiz::findOrFail($id_custom);
+        if (!$customQuiz) throw new CustomNotFound('no quiz found');
         return response()->json($customQuiz);
     }
 
@@ -31,6 +42,7 @@ class CustomQuizController extends Controller
     public function updateCustom(Request $request, $id)
     {
         $customQuiz = CustomQuiz::findOrFail($id);
+        if (!$customQuiz) throw new CustomNotFound('no quiz found');
         $customQuiz->update($request->all());
         $request->validate(CustomQuiz::rulesForUpdate());
         return response()->json($customQuiz, 200);
@@ -40,6 +52,7 @@ class CustomQuizController extends Controller
     public function deleteCustom($id)
     {
         $customQuiz = CustomQuiz::findOrFail($id);
+        if (!$customQuiz) throw new CustomNotFound('no quiz found');
         $customQuiz->delete();
 
         return response()->json(null, 204);
