+++ app/Http/Controllers/UserController.php
@@ -3,8 +3,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\CustomNotFound;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Mockery\Exception;
 
 class UserController extends Controller
 {
@@ -17,19 +20,35 @@ class UserController extends Controller
 
     public function createUser(Request $request)
     {
-        $user = User::create($request->all([
+        $request->validate(User::rulesForUsers());
+        $data = $request->only([
             'username',
             'email',
             'password',
-        ]));
-        $request->validate(User::rulesForUsers());
+        ]);
+
+        $data['password'] = bcrypt($data['password']);
+
+        $user = User::create($data);
+        if (!$user) throw new Exception('error: couldnt create the user', 500);
+
         return response()->json($user, 201);
     }
 
+    public function login(Request $request)
+    {
+        $credentials = $request->only('username', 'password');
+
+        if (Auth::attempt($credentials)) {
+            return response()->json(['message' => 'success login']);
+        }
+        return response()->json(['message' => 'not correct credentials'], 401);
+    }
 
     public function getByUsername($username)
     {
         $user = User::where('username', $username)->firstOrFail();
+        if (!$user) throw new CustomNotFound('user not found');
         return response()->json($user);
     }
 
@@ -37,6 +56,7 @@ class UserController extends Controller
     {
 
         $userSelected = User::where('username', $username)->firstOrFail();
+        if (!$userSelected) throw new CustomNotFound('user not found');
        $request->validate(User::rulesForUpdateUsers());
         $userSelected-> fill($request->only([
             'username',
@@ -47,10 +67,25 @@ class UserController extends Controller
         return response()->json($userSelected);
     }
 
+    public function updateStats(Request $request, $username)
+    {
+
+        $userSelected = User::where('username', $username)->firstOrFail();
+        if (!$userSelected) throw new CustomNotFound('user not found');
+
+        $userSelected-> fill($request->only([
+            'points',
+            'quizzes_resolved'
+        ]));
+        $userSelected->save();
+        return response()->json($userSelected);
+    }
+
 
     public function deleteByUsername($username)
     {
         $user = User::where('username', $username)->firstOrFail();
+        if (!$user) throw new CustomNotFound('user not found');
         $user->delete();
         return response()->json(null, 204);
     }
